chore: tmp rename casing vercel build
