+++ jobs/Frontend/fe-challenge/src/pages/loading-tmp/index.ts
