--- jobs/Frontend/fe-challenge/src/pages/loading/index.ts
@@ -1,3 +1,0 @@
-import Loading from './Loading';
-
-export default Loading;
