+++ jobs/Frontend/fe-challenge/src/pages/movie-detail-tmp/components-tmp/MovieGenres.tsx
