+++ jobs/Frontend/fe-challenge/src/pages/movie-detail-tmp/index.ts
